Fix CSD-MT: use FP32 precision to match original quality
FP16 caused visible artifacts (mean pixel diff 3.24, up to 67).
FP32 reduces to mean 1.49 with near-identical results to PyTorch.
Model size: 53MB + 25MB (was 26MB + 13MB).

+++ conversion_scripts/convert_csdmt.py
@@ -152,7 +152,7 @@ def convert_face_parser():
             ct.TensorType(name="face_mask"),
         ],
         minimum_deployment_target=ct.target.iOS17,
-        compute_precision=ct.precision.FLOAT16,
+        compute_precision=ct.precision.FLOAT32,
     )
 
     ml.author = "CoreML-Models"
@@ -209,7 +209,7 @@ def convert_makeup_transfer():
         ],
         outputs=[ct.TensorType(name="result")],
         minimum_deployment_target=ct.target.iOS17,
-        compute_precision=ct.precision.FLOAT16,
+        compute_precision=ct.precision.FLOAT32,
     )
 
     ml.author = "CoreML-Models"
